Create a second Map from Object to its adjacencies
Before this change, queries with a wildcard for subject would not get
the O(log n) promised in the documentation, with the actual
implementation using multiple O(n) operations.

This patch changes the time/memory tradeoff towards memory, by creating
a secondary map from objects to their adjencies and hooking it up to
query'. This creates redundancy in triples storage (and thereby
increases memory usage), but delivers the promised O(log n)
complexity for queries with either subject or object non-wildcarded.