Switch MGraph from Map, Set to HashMap, HashSet
During use of RDF4h on a large dataset, it became apparent that a lot of
time was spent in the compare function of the Node type. Switching
MGraph over to HashMap and HashSet (i.e. using Hashable instead of Ord to
index the map) seems to be faster for heavier workloads.

Known issue: test suite does not pass for duplicate and select tests.
This is not of interest to my particular use of rdf4h, but should be
fixed if/before this change gets merged.