Sort triples before testing equality in 2 tests
Tests no_dupes and match_none compare lists of triples. Commit
2378635b5357aacecc14b6d0c0fa8b4dfd98ff76 introduces the HashMap
unordered container. So URIs are hashed prior to being added to an RDF
graph. A `triplesOf` call on the RDF graph will likely return the
triples in a different order to the list passed to `mkRdf`. The fix is
to sort triples in the two failing test cases with `sortTriples` before
testing the same triples lists for equality.