Relative/absolute IRI resolution in XML parser
Use the new `IRI` module to resolve IRIs in the XML parser, to
determine whether the base URL is to be prepended to a relative IRI or
not.

This commit means that RDF/XML documents with no triples are parsed to
RDF graphs with no triples, rather than throwing an exception.

The commit increases the number of W3C failures for RDF/XML to 62 out
of 162 tests.