Use xmlbf for XML parsing (#67)
A new implementation of the RDF/XML parser, using the xmlbf library. The previous RDF/XML parser was based on HXT, but failed 65 of the W3C parsing unit tests. This new xmlbf based parser fixes all these failing tests (all but one, being worked on), and is ~25% faster too.