Fixed issue: previously only asserted that the subject existed in the serialization.
Theses tests reveal a possible problem with the `Arbitrary` instance for a `RDF
TList`: the same URI is occassionally used for more than one node type. I.e., a
URI might be used for a subject and predicate or subject and object. The
instance should be updated to not used duplicate URIs for different node types.