Bump Stackage LTS version
Template Haskell 2.18.0 introduces the ability to add Haddock
documentation to declarations that are generated within the `Q` monad.
We tried adding this as an `extra-dep` within the `stack.yaml` file,
but because the Stackage LTS version is so old (over three years!),
and because several of our dependency packages specify incorrect and
unnecessary upper bounds on the Template Haskell version they work
with, doing so would involve pinning many more dependencies than we
already do.  This patch takes the route instead of bumping the
Stackage LTS version to the minimum version that includes Template
Haskell 2.18.